requirements.txt
discord.py
discord.py[voice]
wavelink
python-dotenv
openai
requests
backoff